separate CSV generation into its own class
