+++ charts/DownloadTab.tsx
@@ -1,13 +1,4 @@
-import {
-    extend,
-    sortBy,
-    uniq,
-    flatten,
-    csvEscape,
-    first,
-    last,
-    isMobile
-} from "./Util"
+import { extend, isMobile } from "./Util"
 import * as React from "react"
 import { observable, computed, action } from "mobx"
 import { observer } from "mobx-react"
@@ -16,8 +7,8 @@ import { ChartConfig } from "./ChartConfig"
 import { LoadingIndicator } from "site/client/LoadingIndicator"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faDownload } from "@fortawesome/free-solid-svg-icons/faDownload"
-import { ChartDimensionWithOwidVariable } from "./ChartDimensionWithOwidVariable"
 import classNames from "classnames"
+import { CSVGenerator } from "./CSVGenerator"
 
 interface DownloadTabProps {
     bounds: Bounds
@@ -159,95 +150,6 @@ export class DownloadTab extends React.Component<DownloadTabProps> {
         }
     }
 
-    // Here's where the actual CSV is made
-    @computed get csvBlob() {
-        const { chart } = this.props
-        const yearIsDayVar = chart.yearIsDayVar
-        const dayIndexedCSV = yearIsDayVar ? true : false
-
-        const dimensions = chart.data.filledDimensions.filter(
-            d => d.property !== "color"
-        )
-        const uniqueEntitiesAcrossDimensions =
-            chart.sortedUniqueEntitiesAcrossDimensions
-
-        // only get days if chart has a day-indexed variable, else get years across dimensions
-        const indexingYears = sortBy(
-            dayIndexedCSV
-                ? yearIsDayVar?.yearsUniq
-                : uniq(flatten(dimensions.map(d => d.yearsUniq)))
-        )
-
-        const rows: string[] = []
-
-        const titleRow = ["Entity", "Code", dayIndexedCSV ? "Date" : "Year"]
-
-        dimensions.forEach(dim => {
-            if (this.isFixedYearDimension(dim)) titleRow.push("Year")
-            titleRow.push(csvEscape(dim.fullNameWithUnit))
-        })
-        rows.push(titleRow.join(","))
-
-        uniqueEntitiesAcrossDimensions.forEach(entity => {
-            indexingYears.forEach(year => {
-                const row: (string | number)[] = [
-                    entity,
-                    chart.entityMetaByKey[entity].code ?? "",
-                    chart.formatYearFunction(year)
-                ]
-
-                let rowHasSomeValue = false
-                dimensions.forEach(dim => {
-                    let value = null
-                    if (this.isFixedYearDimension(dim)) {
-                        const latestYearValue = dim.yearAndValueOfLatestValueforEntity(
-                            entity
-                        )
-                        if (latestYearValue) {
-                            row.push(
-                                dim.formatYear(first(latestYearValue) as number)
-                            )
-                            value = last(latestYearValue)
-                        } else row.push("")
-                    } else {
-                        value = dim.valueByEntityAndYear.get(entity)?.get(year)
-                    }
-
-                    if (value != null) {
-                        row.push(value)
-                        rowHasSomeValue = true
-                    } else row.push("")
-                })
-
-                // Only add rows which actually have some data in them
-                if (rowHasSomeValue) rows.push(row.map(csvEscape).join(","))
-            })
-        })
-
-        return new Blob([rows.join("\n")], { type: "text/csv" })
-    }
-
-    @computed get csvDataUri(): string {
-        return window.URL.createObjectURL(this.csvBlob)
-    }
-
-    @computed get csvFilename(): string {
-        return this.props.chart.data.slug + ".csv"
-    }
-
-    // IE11 compatibility
-    @action.bound onDownload(ev: React.MouseEvent<HTMLAnchorElement>) {
-        if (window.navigator.msSaveBlob) {
-            window.navigator.msSaveBlob(this.csvBlob, this.csvFilename)
-            ev.preventDefault()
-        }
-    }
-
-    // returns true if given dimension is year-based in a chart with day-based variable
-    private isFixedYearDimension(dim: ChartDimensionWithOwidVariable) {
-        return this.props.chart.yearIsDayVar && !dim.yearIsDayVar
-    }
-
     renderReady() {
         const {
             props,
@@ -287,6 +189,7 @@ export class DownloadTab extends React.Component<DownloadTabProps> {
         }
 
         const externalCsvLink = this.props.chart.externalCsvLink
+        const csvGenerator = new CSVGenerator({ chart: props.chart })
         const csv_download = (
             <React.Fragment>
                 <div className="download-csv" style={{ maxWidth: "100%" }}>
@@ -296,14 +199,21 @@ export class DownloadTab extends React.Component<DownloadTabProps> {
                     </p>
                     <a
                         href={
-                            externalCsvLink ? externalCsvLink : this.csvDataUri
+                            externalCsvLink
+                                ? externalCsvLink
+                                : csvGenerator.csvDataUri
                         }
-                        download={this.csvFilename}
+                        download={csvGenerator.csvFilename}
                         className="btn btn-primary"
                         data-track-note="chart-download-csv"
-                        onClick={externalCsvLink ? undefined : this.onDownload}
+                        onClick={
+                            externalCsvLink
+                                ? undefined
+                                : csvGenerator.onDownload
+                        }
                     >
-                        <FontAwesomeIcon icon={faDownload} /> {this.csvFilename}
+                        <FontAwesomeIcon icon={faDownload} />{" "}
+                        {csvGenerator.csvFilename}
                     </a>
                 </div>
             </React.Fragment>
